[feature] Comment 예외 코드 추가

+++ src/main/java/study/login/comment/service/CommentServiceImpl.java
@@ -11,26 +11,30 @@ import study.login.comment.domain.Comment;
 import study.login.comment.domain.CommentDto;
 import study.login.comment.domain.CommentListDto;
 import study.login.comment.service.port.CommentRepository;
+import study.login.common.exception.CommentNotFoundException;
+import study.login.common.exception.PermissionDeniedException;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 @Builder
 @RequiredArgsConstructor
-@Transactional
 public class CommentServiceImpl implements CommentService {
 
     private final CommentRepository commentRepository;
     private final ArticleService articleService;
 
+    @Transactional(readOnly = true)
     public List<CommentListDto> findList(Long articleId) {
         return commentRepository.findByArticleId(articleId)
                 .stream().map(CommentListDto::new)
                 .collect(Collectors.toList());
     }
 
+    @Transactional
     public Comment write(CommentDto commentDto) {
         Article article = articleService.findById(commentDto.getArticleId()).orElseThrow(NoSuchElementException::new);
 
@@ -43,11 +47,19 @@ public class CommentServiceImpl implements CommentService {
         return commentRepository.save(comment);
     }
 
+    @Transactional
     public void deleteByArticleId(Long articleId) {
         commentRepository.deleteByArticleId(articleId);
     }
 
-    public void deleteById(Long commentId) {
+    @Transactional
+    public void deleteById(Long commentId, String nickname) {
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(CommentNotFoundException::new);
+
+        if (!comment.getNickname().equals(nickname))
+            throw new PermissionDeniedException();
+
         commentRepository.deleteById(commentId);
     }
 }
